Fixed a casting error in the DispatchCallback method. (was Long, must be Integer)
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@105752 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/jbaci/src/alma/ACS/impl/RWlongImpl.java
@@ -208,9 +208,9 @@ public class RWlongImpl
 		try
 		{	
 			if (type == CallbackDispatcher.DONE_TYPE)
-				((CBlong)callback).done((int) ((Long)value).longValue(), completion, desc);
+				((CBlong)callback).done((int) ((Integer)value).intValue(), completion, desc);
 			else if (type == CallbackDispatcher.WORKING_TYPE)
-				((CBlong)callback).working((int) ((Long)value).longValue(), completion, desc);
+				((CBlong)callback).working((int) ((Integer)value).intValue(), completion, desc);
 			else 
 				return false;
